requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.8
dominate==2.9.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
